+++ Src/Nodes/PropertyNode.cs
@@ -21,21 +21,28 @@ namespace ODataQuery.Nodes
 
   sealed class DatePropertyNode : Node
   {
+    private readonly Node @this;
     private readonly MethodInfo dtProperty;
     private readonly MethodInfo dtoProperty;
-    private readonly Node @this;
+    private readonly MethodInfo doProperty;
 
-    public DatePropertyNode(MethodInfo dtProperty, MethodInfo dtoProperty, Node @this)
+    public DatePropertyNode(Node @this, MethodInfo dtProperty, MethodInfo dtoProperty, MethodInfo doProperty = null)
     {
+      this.@this = @this;
       this.dtProperty = dtProperty;
       this.dtoProperty = dtoProperty;
-      this.@this = @this;
+      this.doProperty = doProperty;
     }
 
     public override Expression ToExpression(Expression instance)
     {
       var target = @this.ToExpression(instance);
-      return Expression.Property(target, target.Type == typeof(DateTime) ? dtProperty : dtoProperty);
+      return Expression.Property(
+        target,
+        target.Type == typeof(DateTime) ? dtProperty
+          : target.Type == typeof(DateTimeOffset) ? dtoProperty
+          : target.Type == typeof(DateOnly) && doProperty != null ? doProperty
+          : throw new InvalidOperationException("Date/time related function cannot be applied to parameter of type " + target.Type));
     }
   }
 }